requirements.txt
albumentations
datasets
fire
manga109api
numpy
opencv_python
opencv_python_headless
pandas
Pillow
torch
torchinfo
transformers
wandb
fugashi
unidic-lite
jiwer
accelerate